+++ Sources/Sentry/PrivateSentrySDKOnly.mm
@@ -135,13 +135,7 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
 
 + (NSString *)installationID
 {
-    NSString *cacheDirectoryPath = self.options.cacheDirectoryPath;
-    if (![SentryOptionsValidator isCacheDirectoryPathValidWithPath:cacheDirectoryPath]) {
-        SENTRY_LOG_FATAL(@"The configured cache directory path looks invalid, the SDK might not be "
-                         @"able to write reports to disk: %@",
-            cacheDirectoryPath);
-    }
-    return [SentryInstallation idWithCacheDirectoryPath:cacheDirectoryPath];
+    return [SentryInstallation idWithCacheDirectoryPath:self.options.cacheDirectoryPath];
 }
 
 + (SentryOptions *)options