+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## Unreleased
 
-### Improvements
-
-- Add error logging for invalid `cacheDirectoryPath` (#4693)
 ### Features
 
 - Add protocol for custom screenName for UIViewControllers (#4646)
+
+### Improvements
+
+- Add error logging for invalid `cacheDirectoryPath` (#4693)
 
 ## 8.43.1-beta.0